output speed optimizations
The main issue appears to have been poor performance concatenating lots of strings to produce multi-hundred-KB (or even multi-MB) output files.  The solution is to make a string for each trackpoint and place those strings in a list that can be concatenated all at once with ''.join().  Performance improvement is _huge_.